@@ -1,6 +1,6 @@
 [package]
 name = "electrs"
-version = "0.9.13"
+version = "0.10.2"
 authors = ["Roman Zeyde <me@romanzey.de>"]
 description = "An efficient re-implementation of Electrum Server in Rust"
 license = "MIT"
@@ -9,7 +9,7 @@ repository = "https://github.com/romanz/electrs"
 keywords = ["bitcoin", "electrum", "server", "index", "database"]
 documentation = "https://docs.rs/electrs/"
 readme = "README.md"
-edition = "2018"
+edition = "2021"
 build = "build.rs"
 
 [features]
@@ -22,37 +22,37 @@ spec = "internal/config_specification.toml"
 
 [dependencies]
 anyhow = "1.0"
-bitcoin = { version = "0.30.0", features = ["serde", "rand-std"] }
+bitcoin = { version = "0.31.0", features = ["serde", "rand-std"] }
+bitcoin_slices = { version = "0.7", features = ["bitcoin", "sha2"] }
+bitcoincore-rpc = { version = "0.18" }
 configure_me = "0.4"
 crossbeam-channel = "0.5"
 dirs-next = "2.0"
-env_logger = "0.9"
+env_logger = "0.10"
 log = "0.4"
-parking_lot = "0.11"
+parking_lot = "0.12"
 prometheus = { version = "0.13", optional = true }
-rayon = "1.5"
+rayon = "1.8"
 serde = "1.0"
-serde_derive = "1.0"
+serde_derive = "1.0, <=1.0.171"  # avoid precompiled binaries (https://github.com/serde-rs/serde/issues/2538)
 serde_json = "1.0"
 signal-hook = "0.3"
 tiny_http = { version = "0.12", optional = true }
 hex_lit = "0.1.1"
 
-[dependencies.electrs-bitcoincore-rpc]
-version = "0.17.0-e3"
-
 [dependencies.electrs-rocksdb]
-# Workaround the following issues:
-# - https://github.com/romanz/electrs/issues/403 (support building on ARM 32-bit)
-# - https://github.com/romanz/electrs/issues/469 (dynamic linking on Debian 11)
-version = "0.15.0-e3"
+version = "0.19.0-e3"
+
 default-features = false
 # ZSTD is used for data compression
 # Snappy is only for checking old DB
 features = ["zstd", "snappy"]
 
 [build-dependencies]
-configure_me_codegen = { version = "0.4.3", default-features = false }
+configure_me_codegen = { version = "0.4.4", default-features = false }
 
 [dev-dependencies]
-tempfile = "3.4"
+bitcoin-test-data = "0.2.0"
+hex_lit = "0.1.1"
+tempfile = "3.8"
+sha2 = "0.10.7"
